DCP/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0

